array access with usize instead of u128.
preparation for changing access to be u64 based.

commit-id:35b266b2

+++ corelib/array.cairo
@@ -4,17 +4,17 @@ extern fn array_append<T>(ref arr: Array::<T>, value: T) nopanic;
 extern fn array_pop_front<T>(ref arr: Array::<T>) -> Option::<T> nopanic;
 #[panic_with('Array out of bounds', array_at)]
 extern fn array_get<T>(
-    ref arr: Array::<T>, index: u128
+    ref arr: Array::<T>, index: usize
 ) -> Option::<T> implicits(RangeCheck) nopanic;
-extern fn array_len<T>(ref arr: Array::<T>) -> u128 nopanic;
+extern fn array_len<T>(ref arr: Array::<T>) -> usize nopanic;
 
 trait ArrayTrait<T> {
     fn new() -> Array::<T>;
     fn append(ref self: Array::<T>, value: T);
     fn pop_front(ref self: Array::<T>) -> Option::<T>;
-    fn get(ref self: Array::<T>, index: u128) -> Option::<T>;
-    fn at(ref self: Array::<T>, index: u128) -> T;
-    fn len(ref self: Array::<T>) -> u128;
+    fn get(ref self: Array::<T>, index: usize) -> Option::<T>;
+    fn at(ref self: Array::<T>, index: usize) -> T;
+    fn len(ref self: Array::<T>) -> usize;
 }
 impl ArrayImpl<T> of ArrayTrait::<T> {
     #[inline(always)]
@@ -30,14 +30,14 @@ impl ArrayImpl<T> of ArrayTrait::<T> {
         array_pop_front(ref self)
     }
     #[inline(always)]
-    fn get(ref self: Array::<T>, index: u128) -> Option::<T> {
+    fn get(ref self: Array::<T>, index: usize) -> Option::<T> {
         array_get(ref self, index)
     }
-    fn at(ref self: Array::<T>, index: u128) -> T {
+    fn at(ref self: Array::<T>, index: usize) -> T {
         array_at(ref self, index)
     }
     #[inline(always)]
-    fn len(ref self: Array::<T>) -> u128 {
+    fn len(ref self: Array::<T>) -> usize {
         array_len(ref self)
     }
 }
